+++ test/TreeTests.cpp
@@ -68,7 +68,7 @@ TEST(TreeTests, FenwickTreeTest)
 	ASSERT_EQ(3, fenwickTree.Query(3)); // f(3) = f(3) + f(2) = 2 + 1 = 3
 	ASSERT_EQ(4, fenwickTree.Query(4)); // f(2)
 }
-class InsertionSortShiftCountTestFixture : public testing::TestWithParam<tuple<long, vector<long>>>
+class InsertionSortShiftCountTestFixture : public testing::TestWithParam<tuple<size_t, vector<long>>>
 {
 public:
 	void SetUp() override
@@ -76,14 +76,14 @@ public:
 		_expected = get<0>(GetParam());
 		_data = get<1>(GetParam());
 	}
-	long InsertionSortShiftCountTest()
+	size_t InsertionSortShiftCountTest()
 	{
 		return _tree.InsertionSortShiftCount(_data);
 	}
 
 protected:
 	FenwickTree<long> _tree;
-	long _expected;
+	size_t _expected;
 	vector<long> _data;
 };
 TEST_P(InsertionSortShiftCountTestFixture, InsertionSortShiftCountTests)
